feat(collections): rename & refactored dialogs into own components

+++ libs/api/collection/data-access/src/lib/collection.service.ts
@@ -51,6 +51,8 @@ export class CollectionService {
           },
         });
 
+        console.log(lastOrderNumber);
+
         return await this.prisma.collection.create({
           data: {
             name,
@@ -111,7 +113,12 @@ export class CollectionService {
   }
 
   async delete(projectId: string, id: string): Promise<void> {
-    await this.getById(id, projectId);
+    const project = await this.getById(id, projectId);
+
+    if (project.name === 'default') {
+      this.logger.error(`Not allowed to delete default collection with ID: ${id}`);
+      throw new InternalServerErrorException('Not allowed to delete default collection');
+    }
 
     try {
       await this.prisma.collection.delete({
